requirements.txt
pandas==2.2.3
google-api-python-client==2.146.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
python-dotenv==1.0.1
